stream_writers refactor
